(refactor) @loginuser추가, user조회 중복로직 제거
