+++ src/main/java/com/techeer/fashioncloud/domain/auth/service/CustomUserDetailService.java
@@ -1,16 +1,12 @@
 package com.techeer.fashioncloud.domain.auth.service;
 
-import com.techeer.fashioncloud.domain.user.entity.User;
+import com.techeer.fashioncloud.domain.auth.util.CustomUser;
 import com.techeer.fashioncloud.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
-
-import java.util.Collections;
 
 @Service
 @RequiredArgsConstructor
@@ -20,18 +16,9 @@ public class CustomUserDetailService implements UserDetailsService {
 
     @Override
     public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        return userRepository.findByEmail(email)
-                .map(this::createUserDetails)
+        CustomUser customUser = userRepository.findByEmail(email)
+                .map(CustomUser::new)
                 .orElseThrow(() -> new UsernameNotFoundException("해당 email을 가진 유저가 존재하지 않습니다 - email: " + email));
-    }
-
-    private UserDetails createUserDetails(User user) {
-        GrantedAuthority authority = new SimpleGrantedAuthority(user.getRole().getRole());
-
-        return new org.springframework.security.core.userdetails.User(
-                user.getEmail(),
-                user.getPassword(),
-                Collections.singleton(authority)
-        );
+        return customUser;
     }
 }
