+++ src/main/java/com/techeer/fashioncloud/domain/auth/util/LoginUser.java
@@ -1,0 +1,14 @@
+package com.techeer.fashioncloud.domain.auth.util;
+
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
+@AuthenticationPrincipal(expression = "#this == 'anonymousUser' ? null : user")
+public @interface LoginUser {
+}
