+++ src/main/java/com/techeer/fashioncloud/global/entity/BaseEntity.java
@@ -1,10 +1,8 @@
 package com.techeer.fashioncloud.global.entity;
 
 
-import jakarta.persistence.Entity;
 import jakarta.persistence.MappedSuperclass;
 import lombok.Getter;
-import lombok.Setter;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 import org.springframework.data.annotation.LastModifiedDate;
@@ -12,7 +10,6 @@ import org.springframework.data.annotation.LastModifiedDate;
 import java.time.LocalDateTime;
 
 @Getter
-@Setter
 @MappedSuperclass
 public class BaseEntity {
     @CreationTimestamp
@@ -21,6 +18,6 @@ public class BaseEntity {
     @UpdateTimestamp
     @LastModifiedDate
     private LocalDateTime updatedAt;
-    //삭제 시점
+
     private LocalDateTime deletedAt;
 }
